+++ sd_card_v2/sd_card_v2.c
@@ -199,7 +199,8 @@ bool sd_fs_finfo(const char *path, fs_file_info_t *finfo)
 	memset(finfo, 0, sizeof(fs_file_info_t));
 
 	// root dir
-	if(strlen(path)<2){
+	if (strlen(path) < 2)
+	{
 		strcpy(finfo->full_name, "/");
 		finfo->is_dir = true;
 		return true;
@@ -225,12 +226,6 @@ fs_file_t *sd_fs_open(const char *file, const char *mode)
 		return NULL;
 	}
 
-	if (!sd_fs_finfo(file, &(fp->file_info)))
-	{
-		free(fp);
-		return NULL;
-	}
-
 	fp->file_ptr = calloc(1, sizeof(FIL));
 	if (!fp->file_ptr)
 	{
@@ -245,13 +240,17 @@ fs_file_t *sd_fs_open(const char *file, const char *mode)
 	}
 	if (strchr(mode, 'w'))
 	{
-		modebyte |= FA_WRITE;
+		modebyte |= FA_CREATE_ALWAYS | FA_WRITE;
 	}
 	if (strchr(mode, 'a'))
 	{
-		modebyte |= FA_OPEN_APPEND;
+		modebyte |= FA_OPEN_APPEND | FA_WRITE;
 	}
 	if (strchr(mode, '+'))
+	{
+		modebyte |= FA_READ | FA_WRITE;
+	}
+	if (strchr(mode, 'x'))
 	{
 		modebyte |= FA_CREATE_NEW;
 	}
@@ -262,6 +261,8 @@ fs_file_t *sd_fs_open(const char *file, const char *mode)
 		free(fp);
 		return NULL;
 	}
+
+	sd_fs_finfo(file, &(fp->file_info));
 
 	return fp;
 }
@@ -594,7 +595,7 @@ CREATE_EVENT_LISTENER_WITHLOCK(grbl_cmd, sd_card_cmd_parser, SD_CARD_BUS_LOCK);
 
 DECL_MODULE(sd_card_v2)
 {
-	//starts the file system and system commands
+	// starts the file system and system commands
 	LOAD_MODULE(file_system);
 	// STARTS SYSTEM MENU MODULE
 	LOAD_MODULE(system_menu);